Adding simplified ds
I confirm that there're no significant difference between .png and .npy in the input pipelines, being more convenient the usage of .npy format to encode the images. The digits were reshaped back to 28x28 after demonstrating that is more convenient to have 2D datasets for visual problems with tensorflow (CNN), when if a simple 1D vector is needed the data can be reshaped using specific layers during training.